output from most recent analysis
So GitHub said I wasn't in sync with the master online repo so I cloned it again, but for some reason knit was giving me an error in a line of code that I'd commented out (so that the script would knit). I'd knitted this before and have run out of time to troubleshoot it for now, but want you to be able to see the output now if you'd like.